refactor: remove ServiceIdentifierOrFunc
