--- src/annotation/lazy_service_identifier.ts
@@ -1,7 +1,0 @@
-import { LazyServiceIdentifier } from '@inversifyjs/common';
-
-import { interfaces } from '../interfaces/interfaces';
-
-export type ServiceIdentifierOrFunc<T> =
-  | interfaces.ServiceIdentifier<T>
-  | LazyServiceIdentifier<T>;
