+++ src/annotation/multi_inject.ts
@@ -1,10 +1,12 @@
+import { LazyServiceIdentifier } from '@inversifyjs/common';
+
 import * as METADATA_KEY from '../constants/metadata_keys';
+import { interfaces } from '../interfaces/interfaces';
 import { DecoratorTarget } from './decorator_utils';
 import { injectBase } from './inject_base';
-import { ServiceIdentifierOrFunc } from './lazy_service_identifier';
 
 const multiInject: <T = unknown>(
-  serviceIdentifier: ServiceIdentifierOrFunc<T>,
+  serviceIdentifier: interfaces.ServiceIdentifier<T> | LazyServiceIdentifier<T>,
 ) => (
   target: DecoratorTarget,
   targetKey?: string | symbol,
